ревью 1 Появилась ошибка при загрузки страници home почему то в js\home.js, в строчке fetch("/user/get20BookDTO/" + currentLang) вместо языка приходит "home"..

+++ src/main/java/com/project/model/ShoppingCartDTO.java
@@ -68,8 +68,8 @@ public class ShoppingCartDTO {
         } else cartItems.addAll(shoppingCartDTO.getCartItems());
     }
 
-    public Integer getTotalCostItems() {
-        Integer cost = 0;
+    public long getTotalCostItems() {
+        long cost = 0;
         for (CartItemDTO cartItemDTO : cartItems) {
             cost += cartItemDTO.getBook().getPrice() * cartItemDTO.getQuantity();
         }
